fixed entry_point.sh on WSL2

+++ bin/entry_point.sh
@@ -2,10 +2,30 @@
 
 CONFIG_FILE=_config.yml 
 
-jekyll_pid=$(pgrep -f jekyll)
-kill -KILL $jekyll_pid
-
 /bin/bash -c "exec jekyll serve --watch --port=8080 --host=0.0.0.0 --livereload --verbose --trace --force_polling"&
+
+SY=$(uname -r)
+
+if [[ $SY = *"WSL"* ]]; then
+
+while true; do
+
+    SUM1=$(md5sum "$CONFIG_FILE")
+    sleep 2
+    SUM2=$(md5sum "$CONFIG_FILE")
+    if [ "$SUM1" != "$SUM2" ];
+    then
+    echo "Change detected to $CONFIG_FILE, restarting Jekyll"
+
+    jekyll_pid=$(pgrep -f jekyll)
+    kill -KILL $jekyll_pid
+
+    /bin/bash -c "rm -f Gemfile.lock && exec jekyll serve --watch --port=8080 --host=0.0.0.0 --livereload --verbose --trace --force_polling"&
+    fi
+
+done
+
+else
 
 while true; do
 
@@ -23,3 +43,5 @@ while true; do
   fi
 
 done
+
+fi
